fastapi_backend/requirements.txt
fastapi[all]>=0.100.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
faster-whisper>=1.0.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
uvicorn>=0.23.0
